Keyboard Interrupts

+++ Siddharth/include/memfn.c
@@ -1,6 +1,6 @@
-#include "system.h"
+#include <system.h>
 
-void *memcpy(void *dest, const void *src, size_t count)
+void *memcpy(void *dest, const void *src, UInt32 count)
 {
     const char *sp = (const char *)src;
     char *dp = (char *)dest;
@@ -8,23 +8,23 @@ void *memcpy(void *dest, const void *src, size_t count)
     return dest;
 }
 
-void *memset(void *dest, char val, size_t count)
+void *memset(void *dest, char val, UInt32 count)
 {
     char *temp = (char *)dest;
     for( ; count != 0; count--) *temp++ = val;
     return dest;
 }
 
-unsigned short *memsetw(unsigned short *dest, unsigned short val, size_t count)
+unsigned short *memsetw(unsigned short *dest, unsigned short val, UInt32 count)
 {
     unsigned short *temp = (unsigned short *)dest;
     for( ; count != 0; count--) *temp++ = val;
     return dest;
 }
 
-size_t strlen(const char *str)
+UInt32 strlen(const char *str)
 {
-    size_t retval;
+    UInt32 retval;
     for(retval = 0; *str != '\0'; str++) retval++;
     return retval;
 }
